All rendering works except for picture and qr code

+++ App/controllers/event.py
@@ -8,6 +8,7 @@ import io
 import base64
 
 
+
 def get_event(event_id):
     return db.session.get(Event, event_id)
 
@@ -23,19 +24,18 @@ def view_event_history(student_id=None, staff_id=None):
         student = db.session.get(Student, student_id)
         return student.attendances
     if staff_id:
-        staff = db.session.get(Staff, staff_id)
-        return staff.events
+        return Event.query.filter_by(staffId=staff_id).all()
     return []
 
 # ---------------- Event CRUD ----------------
 
-def create_event(staff_id, name, type, description, start, end):
-    new_event = Event(name=name, type=type, description=description, start=start, end=end)
-    new_event.staffId = staff_id
-    new_event.qr = generate_qr_code(new_event.id)  # Generate QR code data for the event
+def create_event(staff_id, name, type, description, start, end, location):
+    new_event = Event(staffId=staff_id, name=name, type=type, description=description, start=start, end=end, location=location)
     db.session.add(new_event)
+    db.session.flush()
+    new_event.qr = generate_qr_code(new_event.id)  # Generate QR code data for the event
     db.session.commit()
-    return new_event.get_json()
+    return new_event
 
 def update_event(event_id, **kwargs):
     event = db.session.get(Event, event_id)
@@ -47,7 +47,8 @@ def update_event(event_id, **kwargs):
         'type': 'type',
         'description': 'description',
         'start': 'start',
-        'end': 'end'
+        'end': 'end',
+        'location': 'location'
     }
 
     for key, value in kwargs.items():
@@ -57,12 +58,23 @@ def update_event(event_id, **kwargs):
     db.session.commit()
     return event.get_json()
 
-def delete_event(event_id):
+def delete_event(event_id, staff_id):
     event = db.session.get(Event, event_id)
+
+    print("EVENT:", event)
+    print("EVENT STAFF ID:", event.staffId if event else None)
+    print("JWT STAFF ID:", staff_id)
+
     if not event:
         return False
+
+    if int(event.staffId) != int(staff_id):
+        print("STAFF ID MISMATCH")
+        return False
+
     db.session.delete(event)
     db.session.commit()
+    print("EVENT DELETED FROM DB")
     return True
 
 # ---------------- Student Event Actions ----------------
